first

+++ 17ce/17ce_openwrt_tmp.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
-rm -f /etc/17ce/17ce_openwrt_ins.sh
-wget --no-check-certificate -O /etc/17ce/17ce_openwrt_ins.sh https://raw.githubusercontent.com/erebusx/r-app/17ce/17ce/17ce_openwrt.sh 2>/dev/null  >/dev/null
-chmod +x /etc/17ce/17ce_openwrt_ins.sh
-rm -f /etc/17ce/17ce_openwrt_tmp.sh
+wget --no-check-certificate -O /tmp/17ce_openwrt.sh https://raw.githubusercontent.com/erebusx/r-app/17ce/17ce/17ce_openwrt.sh 2>/dev/null  >/dev/null
+if [ -f "/tmp/17ce_openwrt.sh" ]; then
+  rm -f /etc/17ce/17ce_openwrt_ins.sh
+  chmod +x /etc/17ce/17ce_openwrt_ins.sh
+fi
+rm -f /tmp/17ce_openwrt_tmp.sh
